requirements.txt
pyocr==0.7.2
pillow==8.0.1
wand==0.6.5
pdf2image==1.14.0
opencv-python==4.4.0.46
matplotlib==3.3.3
unicodecsv==0.14.1

